Refactor: Initial project refinement pass.
This commit includes several improvements and cleanups across the project:

1.  **Drizzle Migrations Cleaned:**
    *   I removed redundant/conflicting Drizzle migration files (`0000_sad_wong.sql`, `0001_session.sql`, `0002_games_table.sql`).
    *   I updated `setup_database.sql` to align the `games.downloads` column type (`bigint`) with the Drizzle schema.
    *   I corrected `server/migrate.ts` to point to the `drizzle/` folder for migrations instead of a non-existent `migrations/` folder.

2.  **Gitignore Updated:**
    *   I added the `uploads/` directory to `.gitignore` to prevent user-uploaded content from being tracked.

3.  **Dependency Management:**
    *   I installed numerous missing dependencies identified by `depcheck` (primarily UI-related Radix UI components and other utilities like `nanoid`, `@twurple/api`).
    *   I removed unused dependencies (`@neondatabase/serverless`, `axios`, `jsonwebtoken`, `lodash`, `uuid`) and devDependencies (`@types/jsonwebtoken`, `@types/lodash`, `@types/uuid`, `dotenv-cli`) after verification.

4.  **Database Management Scripts Consolidated:**
    *   I removed the redundant `server/seed-games.ts` script.
    *   I renamed `server/init-games.ts` to `server/setup-games.ts`.
    *   I removed migration execution responsibility from `server/setup-games.ts` to centralize it under `npm run db:migrate`.
    *   I updated `package.json` scripts (`db:seed`, `init-games`) to reflect the new workflow (migrate -> seed users -> seed games).

5.  **Code Refinements (Server-side):**
    *   I improved the main error handler in `server/index.ts` to send generic messages in production for security, while retaining detailed server-side logs. The error response format is now `{ error: { message: "..." } }`.
    *   I removed the basic/redundant `server/db.test.ts` script.
    *   I enhanced `server/check-users.ts` to display password status (Hashed, Not Hashed, Empty) instead of partial password hashes, improving security for this diagnostic tool.

**Next Steps (not included in this commit):**
*   Further review of client-side code for bugs and best practices.
*   Implementation of tests, starting with API endpoint tests (e.g., user login). A plan for a script-based login test was formulated.
*   Addressing potential route definition overlaps for user management endpoints.

+++ setup_database.sql
@@ -4,7 +4,7 @@ CREATE TABLE "games" (
 	"categories" json NOT NULL,
 	"platforms" json NOT NULL,
 	"contact" text,
-	"downloads" integer,
+	"downloads" bigint,
 	CONSTRAINT "games_name_unique" UNIQUE("name")
 );
